requirements.txt
pandas==2.2.3
numpy==1.26.3
tqdm==4.67.1

torch==2.5.1+cu121
transformers==4.48.2
tokenizers==0.21.0

openai==1.74.0
google-genai==0.5.0

rouge-score==0.1.2
datasets==3.2.0
